feat(core): add DECLRMM mode 69 plumbing (§08.3)
Widen TermMode from u32 to u64 to accommodate bit 32, then add the
full vertical slice for DEC Left Right Margin Mode (DECLRMM):

- VTE: NamedPrivateMode::LeftRightMargin = 69 variant + PrivateMode::new mapping
- TermMode: LEFT_RIGHT_MARGIN = 1 << 32 flag
- helpers: named_private_mode_flag exhaustive match arm
- modes: DECSET/DECRST handler arms (insert/remove flag)
- Mode reporting: DECRQM works automatically via the flag mapping

Cascade: AtomicU32→AtomicU64 for mode_cache across oriterm_mux pane,
io_thread, domain; wire protocol snapshot.modes u32→u64; MuxBackend
trait pane_mode returns u64; all test call sites updated.

Tests: mode_69_set, mode_69_reset, mode_69_decrqm_reports, exhaustive
decset_decrst_flag_sync updated, all_flags_are_distinct updated,
term_mode_size guard updated to 8 bytes.

+++ oriterm_mux/src/backend/client/rpc_methods.rs
@@ -107,7 +107,7 @@ impl MuxBackend for MuxClient {
         self.dirty_panes.insert(pane_id);
     }
 
-    fn pane_mode(&self, pane_id: PaneId) -> Option<u32> {
+    fn pane_mode(&self, pane_id: PaneId) -> Option<u64> {
         self.pane_snapshots.get(&pane_id).map(|s| s.modes)
     }
